# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API service for the District Core Crime Management System",
        "title": "District Core Crime Management System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/cases": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a paginated list of cases with optional search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "List all cases",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term for case name or description",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cases and total count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new criminal case record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Create a new case",
                "parameters": [
                    {
                        "description": "Case details",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.CaseDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Case"
                        }
                    },
                    "400": {
                        "description": "Invalid case data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cases/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detailed information about a specific case",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Get case details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Case"
                        }
                    },
                    "400": {
                        "description": "Invalid case ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Case not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing criminal case record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases"
                ],
                "summary": "Update an existing case",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated case details",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.CaseDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Case"
                        }
                    },
                    "400": {
                        "description": "Invalid case data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Permission denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Case not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/cases/{id}/assignees": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve list of users assigned to a case",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cases",
                    "assignees"
                ],
                "summary": "Get case assignees",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid case ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/public/reports": {
            "post": {
                "description": "Public endpoint to submit a crime report",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public",
                    "reports"
                ],
                "summary": "Submit a crime report",
                "parameters": [
                    {
                        "description": "Crime report details",
                        "name": "report",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_dto.ReportDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "reportId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid report data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/public/reports/{reportId}/status": {
            "get": {
                "description": "Public endpoint to check a case status using report ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "public",
                    "reports"
                ],
                "summary": "Get case status by report ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report ID",
                        "name": "reportId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report ID and status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Report ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Report not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_m7medVision_crime-management-system_internal_dto.CaseDTO": {
            "type": "object",
            "required": [
                "area",
                "authorizationLevel",
                "caseType",
                "description",
                "name",
                "reportedById"
            ],
            "properties": {
                "area": {
                    "type": "string"
                },
                "authorizationLevel": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.ClearanceLevel"
                },
                "caseType": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "reportedById": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_dto.ReportDTO": {
            "type": "object",
            "required": [
                "description",
                "location",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.Case": {
            "type": "object",
            "properties": {
                "area": {
                    "description": "City/Area",
                    "type": "string"
                },
                "assignees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                    }
                },
                "authorizationLevel": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.ClearanceLevel"
                },
                "caseNumber": {
                    "type": "string"
                },
                "caseType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                },
                "createdByID": {
                    "type": "integer"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "evidence": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Evidence"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "reportedBy": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                },
                "reportedByID": {
                    "type": "integer"
                },
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Report"
                    }
                },
                "status": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.CaseStatus"
                },
                "suspects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Suspect"
                    }
                },
                "updatedAt": {
                    "type": "string"
                },
                "victims": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Victim"
                    }
                },
                "witnesses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Witness"
                    }
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.ClearanceLevel": {
            "type": "string",
            "enum": [
                "low",
                "medium",
                "high",
                "critical"
            ],
            "x-enum-varnames": [
                "ClearanceLow",
                "ClearanceMedium",
                "ClearanceHigh",
                "ClearanceCritical"
            ]
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.Report": {
            "type": "object",
            "properties": {
                "cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Case"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "reportID": {
                    "type": "string"
                },
                "reportedBy": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.User"
                },
                "reportedByID": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_m7medVision_crime-management-system_internal_model.User": {
            "type": "object",
            "properties": {
                "clearanceLevel": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.ClearanceLevel"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastLogin": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_m7medVision_crime-management-system_internal_model.Role"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}